requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
matplotlib==3.7.1
pandas==1.5.3
numpy==1.24.3  # Явно указываем совместимую версию
psycopg2-binary==2.9.6
python-multipart==0.0.6